[FIX] fix test_assets_bundle_css_error_frontend
Test was failing as website module is not dependency
for test_assetsbundle, but we were trying to call website model.
As a solution, test was moved from test_assetsbundle to website.

X-original-commit: 8c5b0e2fc738cd4666ff1489bbf7c4d98f6256fa
