+++ addons/website/static/tests/tours/test_css_error.js
@@ -2,9 +2,9 @@
 
 import { registry } from "@web/core/registry";
 
-registry.category("web_tour.tours").add('css_error_tour', {
+registry.category("web_tour.tours").add('css_error_tour_frontend', {
     test: true,
-    url: '/web',
+    url: '/',
     steps: () => [
     {
         content: "Error message",
